# openapi.yaml
openapi: 3.0.3
info:
  title: ias_stt — Trap Management API
  version: 1.0.0
  description: |
    Trap management service with deployment tracking, location
    history, and photo upload. Telemetry is ingested from the
    ChirpStack HTTP integration via `POST /api/telemetry/ingest`
    (an optional MQTT listener remains available). Application users obtain
    JWT access and refresh tokens from `/auth/login` and send access tokens in
    the `Authorization` header. A legacy API key remains available for service
    integrations.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate a local user
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        200:
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      operationId: currentUser
      responses:
        200:
          description: Current user and permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        401:
          $ref: '#/components/responses/401'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for an access token
      operationId: refreshToken
      responses:
        200:
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, example: Bearer }
                  expires_in: { type: integer, example: 900 }
        401:
          $ref: '#/components/responses/401'
...
  /api/auth/verify:
    get:
      tags: [Auth]
      summary: Verify API key
      operationId: verifyApiKey
      responses:
        200:
          description: Key is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
              example:
                ok: true
        401:
          $ref: '#/components/responses/401'

  # -- Telemetry -------------------------------------------------------------
  /api/telemetry/ingest:
    post:
      tags: [Telemetry]
      summary: Ingest a decoded GPS tracker payload
      description: |
        Receives the ChirpStack JSON uplink envelope pushed by the
        ChirpStack HTTP integration. Processing is identical to the
        (optional) MQTT path: `deviceInfo.devEui` is looked up in the
        trackers table; unknown devices are silently ignored (still
        `200`). `object.latitude`, `object.longitude`,
        `object.position`, and `object.battery` update the tracker,
        and the geofence/deployment logic runs when coordinates are
        present.
      operationId: telemetryIngest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryPayload'
            example:
              deviceInfo:
                devEui: a1b2c3d4e5f6a7b8
                deviceName: trap-tracker-01
              object:
                latitude: -1.2921
                longitude: 36.8219
                position: normal
                battery: 95
      responses:
        200:
          description: Payload accepted and processed (unknown devEui included)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [processed] }
              example:
                status: processed
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'

  # -- Uplink history ---------------------------------------------------------
...
  /api/traps:
    get:
      tags: [Traps]
      summary: List traps
      operationId: listTraps
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive] }
          description: Filter by trap status
      responses:
        200:
          description: A JSON array of traps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trap'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'

    post:
      tags: [Traps]
      summary: Create a trap
      operationId: createTrap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrapCreate'
            example:
              status: active
              trap_id: TRAP-001
              tracker_id: TRK-001
              location: north
              temperature: 23.5
              updated_by: alice
      responses:
        201:
          description: Trap created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trap'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        409:
          $ref: '#/components/responses/409'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Application users pass the JWT access token as
        `Authorization: Bearer <access-token>`. Legacy service integrations
        may pass the configured API key in the same header.

  # -- Reusable parameters ---------------------------------------------------
  parameters:
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 100, minimum: 0 }
      description: Maximum items per page
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
      description: Number of items to skip

  # -- Schemas ---------------------------------------------------------------
  schemas:
    Error:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code: { type: string }
                message: { type: string }
      example:
        error:
          code: AUTHENTICATION_REQUIRED
          message: Authentication is required.
    AuthUser:
      type: object
      properties:
        username: { type: string }
        display_name: { type: string, nullable: true }
        email: { type: string, nullable: true }
        role: { type: string, enum: [administrator, field_operator, read_only] }
        permissions:
          type: array
          items: { type: string }
    LoginResponse:
      type: object
      properties:
        authenticated: { type: boolean }
        user: { $ref: '#/components/schemas/AuthUser' }
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, example: 900 }
        refresh_expires_in: { type: integer, example: 604800 }
        expires_at: { type: string, format: date-time }
        refresh_expires_at: { type: string, format: date-time }
    AuthMeResponse:
      type: object
      properties:
        authenticated: { type: boolean }
        user: { $ref: '#/components/schemas/AuthUser' }
        expires_at: { type: string, format: date-time }
...
    TelemetryPayload:
      type: object
      description: ChirpStack decoded uplink envelope.
      properties:
        deviceInfo:
          type: object
          properties:
            devEui:      { type: string, description: LoRaWAN device EUI; must match a tracker row. }
            deviceName:  { type: string }
        object:
          type: object
          properties:
            latitude:    { type: number, format: float }
            longitude:   { type: number, format: float }
            position:    { type: string, description: Mapped to tracker tilt_status. }
            battery:     { type: integer }
      required: [deviceInfo]
...
    Trap:
      type: object
      properties:
        id:                { type: integer, readOnly: true }
        status:            { type: string, enum: [active, inactive] }
        trap_id:           { type: string, maxLength: 50 }
        tracker_id:        { type: string, maxLength: 50 }
        location:          { type: string, maxLength: 50, nullable: true }
        door_status:       { type: string, maxLength: 20, nullable: true }
        temperature:       { type: number, format: float, nullable: true }
        notes:             { type: string, maxLength: 255, nullable: true }
        updated_by:        { type: string, maxLength: 50, description: Ignored; derived from the authenticated principal. }
        created_at:        { type: string, format: date-time, readOnly: true }
        updated_at:        { type: string, format: date-time, readOnly: true }
      required:
        - status
        - trap_id
    TrapCreate:
      type: object
      properties:
        status:            { type: string, enum: [active, inactive] }
        trap_id:           { type: string, maxLength: 50 }
        tracker_id:        { type: string, maxLength: 50, description: Device EUI of an existing tracker; may be omitted/empty. }
        location:          { type: string, maxLength: 50 }
        door_status:       { type: string, maxLength: 20 }
        temperature:       { type: number, format: float }
        notes:             { type: string, maxLength: 255 }
        updated_by:        { type: string, maxLength: 50, description: Ignored; derived from the authenticated principal. }
      required: [status, trap_id]
...
  responses:
    400:
      description: Bad request — missing or invalid fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingField:
              summary: Missing required field
              value: { error: "Missing required field(s): status" }
            invalidType:
              summary: Invalid limit/offset
              value: { error: "'limit' and 'offset' must be integers" }
    401:
      description: Missing or invalid JWT or service credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Authentication is required.
    403:
      description: Authenticated user lacks the required permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action.
...
    409:
      description: Conflict — duplicate trap_id
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: trap_id 'TRAP-001' already exists
...